SynQuest/SynQuest_v1.0.0/requirements.txt
textual>=0.52.0
openai>=1.13.0
python-dotenv>=1.0.1
pydantic>=2.6.2
